requirements.txt
google-cloud-secret-manager==2.19
openai==1.16.2
google-cloud-storage==2.16.0
Flask==3.0.2
Pillow==10.3.0

